+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-02-11 18:37:24
+
+### Changed Files
+- `supabase/migrations/20260211161810_fix_final_submit_dates.sql`
+
+### Statistics
+```
+ 1 file changed, 19 insertions(+)
+```
+
 ## [Checkpoint] - 2026-02-11 14:25:45
 
 ### Changed Files